Changing js files to tsx files
